+++ src/test/java/io/wzw/backend/data/dao/impl/HibernateMeetupDAOTest.java
@@ -39,7 +39,7 @@ public class HibernateMeetupDAOTest {
 	@Test
 	public void testSelectById() {
 		// Select after insert
-		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1, 19.015f, 20.013f, new User());
+		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1L, 19.015f, 20.013f, new User());
 				meetupDAO.insert(insertMeetup);
 		Meetup meetup = meetupDAO.selectById(insertMeetup.getId(), Meetup.class);
 		assertEquals("Select by Id should exist",meetup.getId(), insertMeetup.getId());
@@ -52,7 +52,7 @@ public class HibernateMeetupDAOTest {
 	@Test
 	public void testSelectAll() {
 		int totalElements = meetupDAO.selectAll(Meetup.class).size();
-		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1, 19.015f, 20.013f, new User());
+		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1L, 19.015f, 20.013f, new User());
 		meetupDAO.insert(insertMeetup);
 		int totalElementsAfterInsert = meetupDAO.selectAll(Meetup.class).size();
 		
@@ -64,7 +64,7 @@ public class HibernateMeetupDAOTest {
 	 */
 	@Test
 	public void testInsert() {
-		Meetup insertMeetup =  new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1, 19.015f, 20.013f, new User());
+		Meetup insertMeetup =  new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1L, 19.015f, 20.013f, new User());
 		meetupDAO.insert(insertMeetup);
 		
 		Meetup meetup = meetupDAO.selectById(insertMeetup.getId(), Meetup.class);
@@ -80,7 +80,7 @@ public class HibernateMeetupDAOTest {
 		String updatedName = "Admin changed";
 		
 		// Select after first insert
-		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1, 19.015f, 20.013f, new User());
+		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1L, 19.015f, 20.013f, new User());
 		meetupDAO.insert(insertMeetup);
 		
 		// We update the role
@@ -99,7 +99,7 @@ public class HibernateMeetupDAOTest {
 	@Test
 	public void testDelete() {
 		// Select after first insert
-		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1, 19.015f, 20.013f, new User());
+		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1L, 19.015f, 20.013f, new User());
 		meetupDAO.insert(insertMeetup);
 		
 		// Delete 
@@ -112,12 +112,12 @@ public class HibernateMeetupDAOTest {
 	
 	@Test
 	public void testGetUserMeetupById() {
-		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1, 19.015f, 20.013f, new User());
+		Meetup insertMeetup = new Meetup("Admin","Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1L, 19.015f, 20.013f, new User());
 		User user = new User(null, "mikel", "1234", "mikel@mikel.com", new Avatar(), new HashSet<Role>(), new ArrayList<Meetup>());
 		insertMeetup.setUser(user);
 		user.getMeetups().add(insertMeetup);
 		meetupDAO.insert(insertMeetup);
-		Meetup meetup = meetupDAO.getUserMeetupById(user.getId().longValue(), insertMeetup.getId().longValue());
+		Meetup meetup = meetupDAO.getUserMeetupById(user.getId(), insertMeetup.getId());
 		assertTrue(meetup.getId() == insertMeetup.getId() && meetup.getUser().getId() == user.getId());
 	}
 	
@@ -127,12 +127,12 @@ public class HibernateMeetupDAOTest {
 		ArrayList<Meetup> meets = new ArrayList<Meetup>();
 		
 		for (int i = 0; i<20; i++){
-			meets.add(new Meetup("Admin"+i,"Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1, 19.015f, 20.013f, user));
+			meets.add(new Meetup("Admin"+i,"Administrator role", new Date(1992, 04, 19), new Date(1992, 04, 12), 1L, 19.015f, 20.013f, user));
 		}
 		user.setMeetups(meets);
 		HibernateUserDAO userDAO = new HibernateUserDAO();
 		userDAO.insert(user);
-		ArrayList<Meetup> result = (ArrayList<Meetup>) meetupDAO.lastEvents(user.getId().longValue(), meets.get(meets.size() - 1).getId().longValue());
+		ArrayList<Meetup> result = (ArrayList<Meetup>) meetupDAO.lastEvents(user.getId(), meets.get(meets.size() - 1).getId());
 		
 		assertTrue(result.size() == 10);
 	}
